+++ app/views/admin/users/controls/_export.html.erb
@@ -1,3 +1,3 @@
-<button type="button" class="js-export-csv-modal-trigger c-btn" data-modal-trigger data-modal-title="Download member data" aria-label="Download member data" data-modal-size="small" data-modal-content-selector=".js-export-csv-modal">
+<button type="button" class="js-export-csv-modal-trigger c-btn" data-modal-title="Download member data" aria-label="Download member data" data-modal-size="small" data-modal-content-selector=".js-export-csv-modal">
   <%= crayons_icon_tag(:export, class: "c-btn__icon", aria_hidden: true) %>
 </button>
