+++ app/views/admin/users/show/overview/_credits.html.erb
@@ -8,7 +8,6 @@
 <div class="flex justify-between items-center">
   <p data-testid="user-credits" class="fs-2xl fw-heavy"><%= @user.unspent_credits_count %></p>
   <button class="c-btn c-btn--secondary whitespace-nowrap" type="button"
-    data-modal-trigger
     data-modal-title="Adjust balance"
     data-modal-size="small"
     data-modal-content-selector=".js-adjust-balance"
