+++ app/views/admin/users/show/profile/_actions.html.erb
@@ -10,13 +10,12 @@
         <% if @user.access_locked? %>
           <li><%= link_to "Unlock access", unlock_access_admin_user_path(@user), method: :patch, class: "c-link c-link--block" %></li>
         <% end %>
-        <li><button type="button" class="c-btn w-100 align-left" data-modal-trigger data-modal-title="Export <%= @user.name %>'s data" data-modal-size="small" data-modal-content-selector=".js-export-data">Export data</button></li>
-        <li><button type="button" class="c-btn w-100 align-left" data-modal-trigger data-modal-title="Merge users" data-modal-size="small" data-modal-content-selector=".js-merge-accounts">Merge users</button></li>
+        <li><button type="button" class="c-btn w-100 align-left" data-modal-title="Export <%= @user.name %>'s data" data-modal-size="small" data-modal-content-selector=".js-export-data">Export data</button></li>
+        <li><button type="button" class="c-btn w-100 align-left" data-modal-title="Merge users" data-modal-size="small" data-modal-content-selector=".js-merge-accounts">Merge users</button></li>
         <% if @user.articles_count > 0 %>
           <li>
             <button type="button"
               class="c-btn w-100 align-left"
-              data-modal-trigger
               data-modal-title="Unpublish all posts"
               data-modal-size="small"
               data-modal-content-selector=".js-unpublish-all-posts"
@@ -25,12 +24,11 @@
           </li>
         <% end %>
         <% if @user.identities.any? %>
-          <li><button type="button" class="c-btn w-100 align-left" data-modal-trigger data-modal-title="Remove social accounts" data-modal-size="medium" data-modal-content-selector=".js-remove-social-accounts">Remove social accounts</button></li>
+          <li><button type="button" class="c-btn w-100 align-left" data-modal-title="Remove social accounts" data-modal-size="medium" data-modal-content-selector=".js-remove-social-accounts">Remove social accounts</button></li>
         <% end %>
         <li>
           <button type="button"
             class="c-btn c-btn--destructive w-100 align-left"
-            data-modal-trigger
             data-modal-title="Banish <%= @user.name %>"
             data-modal-size="small"
             data-modal-content-selector=".js-banish-for-spam"
@@ -38,7 +36,7 @@
             data-user-name="<%= @user.name %>"
             data-banishable-user="<%= @banishable_user %>">Banish user</button>
         </li>
-        <li><button type="button" class="c-btn c-btn--destructive w-100 align-left" data-modal-trigger data-modal-title="Delete <%= @user.name %>'s account" data-modal-size="small" data-modal-content-selector=".js-delete-user">Delete user</button></li>
+        <li><button type="button" class="c-btn c-btn--destructive w-100 align-left" data-modal-title="Delete <%= @user.name %>'s account" data-modal-size="small" data-modal-content-selector=".js-delete-user">Delete user</button></li>
       </ul>
     </div>
   </div>
